requirements.txt
flask==2.3.3        
google-cloud-storage==2.19.0  
tensorflow==2.17.1
numpy==1.24.4
